Actual confirmed bugfix for VLAN parsing issue.
Originally ngrep only accounted for VLAN offset if it was
specified in the BPF filter.

Prior version 1.46 expanded the BPF filter to include IPv4/6
traffic, regardless of whether or not VLANs were in use -- but
the offset calculation didn't account for the variable presence
of VLAN headers.

This version keeps the inclusive BPF filter, but detects VLAN
frames in every packet and adjusts the offset on the fly.