Fix pcap datalink bug introducd with VLAN changes
Context: a bug report came in while ago on ngrep for not "seeing" TCP
frames in a pcap dump, when there obviously were TCP frames present.
That ended up being due to the fact that the PCAP data had been captured
off a VLAN, and thus had VLAN headers embedded in the frames.  Ended up
not being ngrep's fault per se; all the user had to do was specify
`vlan` in their BPF filter and PCAP would deliver the frames to ngrep.
But for a netops person moving across lots of different interfaces and
pcap files quickly, it's hard to keep track and you kind of just want it
to work.  Favoring convenience at the time, I thought it would be both
helpful and harmless to include VLAN frames automatically in the BPF
filter.

Turns out it wasn't harmless.  I found newer bug reports for "no VLAN
support for data link type 113", which is the PCAP error message you get
when you combine `vlan` in your BPF filter with frames that come from a
Linux "cooked" socket (what you get with `-d any`).

I'm still working out what to do in the long run, around Just Working
when VLANs are in the mix.  In the meantime, this reworks how the
filters are constructed, introduces the `include_vlan` flag, switches it
off when `-d any` is specified, and moves BPF filter construction to
later in the initialization so there's more opportunity to conditionally
switch it off (e.g. when reading from a `DLT_LINUX_SLL` device).