Always count matches, and emit total upon exit
For long-running sessions the number will wrap back to 0, but I see no
harm in that vs. trying to make a 64bit number work across all the
supported platforms.  They all overflow in the end.

Also no longer emits pcap_stats on exit; pcap_stat manpage basically
says the entire structure is unreliable across platforms.  "May or may
not be useful", is what it should say.  Instead we emit ngrep's stats,
so as not to cause confusion by visibly exiting with a consistent
mismatch between ngrep vs. PCAP stats.

Credit to @rwhalb in #5 for the suggestion.