Add stream following capability
The -J option will print packets following a match but only those that
are in the same TCP stream as the match.

If another packet matches during the following, the number counter is
reset. This helps mitigate problems with a match not having enough
packets after it in the same stream (which could mean ngrep prints
nothing forever after the match). This functionality can be defeated
with the -Q option.