Add in err.h header to tcpkill
Problem I've had with it is that I don't know if that header is present
on all the other OSes ngrep builds on, and I don't have access to them
like back in the day..

OTOH, I don't see a strong reason to hold back on good/correct when
most users are on Linux or macOS these days.

So, if this change breaks the build on your OS, *please* let me know
and I'll fix it.